✨ Add tracking validator's missed block counter, moniker, status
